+++ client/src/app/(landing)/layout-providers.tsx
@@ -1,22 +1,16 @@
 'use client';
 
-import { PropsWithChildren, useCallback, useState } from 'react';
+import { PropsWithChildren, useCallback } from 'react';
 
 import { MapProvider } from 'react-map-gl';
 
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { RecoilRoot } from 'recoil';
 
 import { RecoilURLSyncNext } from '@/lib/recoil';
 import type { Deserialize, Serialize } from '@/lib/recoil';
 import RecoilDevTools from '@/lib/recoil/devtools';
 
-import { TooltipProvider } from '@/components/ui/tooltip';
-
-import { notesESA, sans } from '@/styles/fonts';
-
 export default function Providers({ children }: PropsWithChildren) {
-  const [queryClient] = useState(() => new QueryClient());
   const serialize: Serialize = useCallback((x) => {
     return x === undefined ? '' : JSON.stringify(x);
   }, []);
@@ -28,29 +22,18 @@ export default function Providers({ children }: PropsWithChildren) {
 
   return (
     <>
-      <style jsx global>{`
-        :root {
-          --font-inter: ${sans.style.fontFamily};
-          --font-esa-notes: ${notesESA.style.fontFamily};
-        }
-      `}</style>
-
-      <QueryClientProvider client={queryClient}>
+      <MapProvider>
         <RecoilRoot>
-          <MapProvider>
-            <TooltipProvider>
-              <RecoilURLSyncNext
-                location={{ part: 'queryParams' }}
-                serialize={serialize}
-                deserialize={deserialize}
-              >
-                <RecoilDevTools />
-                {children}
-              </RecoilURLSyncNext>
-            </TooltipProvider>
-          </MapProvider>
+          <RecoilURLSyncNext
+            location={{ part: 'queryParams' }}
+            serialize={serialize}
+            deserialize={deserialize}
+          >
+            <RecoilDevTools />
+            {children}
+          </RecoilURLSyncNext>
         </RecoilRoot>
-      </QueryClientProvider>
+      </MapProvider>
     </>
   );
 }
