+++ client/src/app/(landing)/layout.tsx
@@ -5,23 +5,12 @@ import Providers from '@/app/(landing)/layout-providers';
 
 import Map from '@/containers/map';
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default function LandingLayout({ children }: { children: React.ReactNode }) {
   return (
     <Providers>
-      <html lang="en">
-        <body className="font-notes h-screen w-screen">
-          <main>
-            <Map />
+      <Map />
 
-            {children}
-          </main>
-        </body>
-      </html>
+      {children}
     </Providers>
   );
 }
