+++ client/src/app/(landing)/page.tsx
@@ -1,5 +1,10 @@
 import Home from '@/containers/home';
 
+export const metadata = {
+  title: 'Impact sphere',
+  description: 'Impact sphere',
+};
+
 export default async function HomePage() {
   return <Home />;
 }
