+++ client/src/app/(static)/layout.tsx
@@ -1,21 +1,8 @@
 import '@/styles/globals.css';
 import '@/styles/mapbox.css';
 
-import Providers from '@/app/(landing)/layout-providers';
+import Providers from '@/app/(static)/layout-providers';
 
-export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <Providers>
-      <html lang="en">
-        <body className="font-notes h-screen w-screen">
-          <main>{children}</main>
-        </body>
-      </html>
-    </Providers>
-  );
+export default function StaticLayout({ children }: { children: React.ReactNode }) {
+  return <Providers>{children}</Providers>;
 }
