+++ client/src/app/layout-providers.tsx
@@ -1,0 +1,28 @@
+'use client';
+
+import { PropsWithChildren, useState } from 'react';
+
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+import { TooltipProvider } from '@/components/ui/tooltip';
+
+import { notesESA, sans } from '@/styles/fonts';
+
+export default function Providers({ children }: PropsWithChildren) {
+  const [queryClient] = useState(() => new QueryClient());
+
+  return (
+    <>
+      <style jsx global>{`
+        :root {
+          --font-inter: ${sans.style.fontFamily};
+          --font-esa-notes: ${notesESA.style.fontFamily};
+        }
+      `}</style>
+
+      <QueryClientProvider client={queryClient}>
+        <TooltipProvider>{children}</TooltipProvider>
+      </QueryClientProvider>
+    </>
+  );
+}
